refactor(chain): reorganize poll module structure and separate vote logic
