+++ packages/chain/src/runtime/modules/poll/index.ts
@@ -1,0 +1,2 @@
+export * from "./vote";
+export * from "./poll";
